# data-processing-service/contracts/data-processing-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Green AI - Data Processing Service",
    "description": "Históricos de Monitoring y Supabase separados; preparación provisional para Prediction. Servicio interno detrás del Gateway.",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/v1/historical-logs": {
      "get": {
        "tags": [
          "Supabase"
        ],
        "summary": "Get Historical Logs",
        "description": "Solo lectura de logs. Máximo 31 días; UTC conserva fracciones. Seguir nextCursor hasta página vacía.\n\nhardwareId es un filtro, no autorización por usuario. Publicar únicamente detrás del Gateway autorizado.",
        "operationId": "get_historical_logs_api_v1_historical_logs_get",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "End"
            }
          },
          {
            "name": "hardwareId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 20,
              "title": "Hardwareid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 2048
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoricalPage"
                }
              }
            }
          },
          "422": {
            "description": "Consulta inválida o límite excedido",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "requestId": {
                      "title": "Requestid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "code",
                    "requestId"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "502": {
            "description": "Respuesta upstream inválida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "requestId": {
                      "title": "Requestid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "code",
                    "requestId"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "description": "Fuente ausente, sin permisos o indisponible",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "requestId": {
                      "title": "Requestid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "code",
                    "requestId"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "504": {
            "description": "Timeout de la fuente",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "requestId": {
                      "title": "Requestid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "code",
                    "requestId"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "requestId": {
                      "title": "Requestid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "code",
                    "requestId"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prediction/dataset": {
      "get": {
        "tags": [
          "Prediction (provisional)"
        ],
        "summary": "Get Prediction Dataset",
        "description": "Prepara CPU (%) de un recurso desde Monitoring. Contrato provisional; no ejecuta inferencia.\n\nSe excluyen missing/non_finite; no se mezclan orígenes ni se imputan huecos.",
        "operationId": "get_prediction_dataset_api_v1_prediction_dataset_get",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "End"
            }
          },
          {
            "name": "cluster",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,127}$",
              "title": "Cluster"
            }
          },
          {
            "name": "resourceId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,127}$",
              "title": "Resourceid"
            }
          },
          {
            "name": "stepSeconds",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 3600,
              "minimum": 15,
              "default": 15,
              "title": "Stepseconds"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionDataset"
                }
              }
            }
          },
          "422": {
            "description": "Consulta inválida o límite excedido",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "requestId": {
                      "title": "Requestid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "code",
                    "requestId"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "502": {
            "description": "Respuesta upstream inválida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "requestId": {
                      "title": "Requestid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "code",
                    "requestId"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "description": "Fuente ausente, sin permisos o indisponible",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "requestId": {
                      "title": "Requestid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "code",
                    "requestId"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "504": {
            "description": "Timeout de la fuente",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "requestId": {
                      "title": "Requestid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "code",
                    "requestId"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "requestId": {
                      "title": "Requestid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "code",
                    "requestId"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HistoricalPage": {
        "properties": {
          "schemaVersion": {
            "type": "string",
            "const": "1.0",
            "title": "Schemaversion"
          },
          "hardwareId": {
            "type": "string",
            "title": "Hardwareid"
          },
          "period": {
            "$ref": "#/components/schemas/Period"
          },
          "units": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Units"
          },
          "dataStatus": {
            "type": "string",
            "enum": [
              "complete",
              "partial",
              "no_data"
            ],
            "title": "Datastatus"
          },
          "recordCount": {
            "type": "integer",
            "title": "Recordcount"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/HistoricalRecord"
            },
            "type": "array",
            "title": "Records"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "schemaVersion",
          "hardwareId",
          "period",
          "units",
          "dataStatus",
          "recordCount",
          "records",
          "nextCursor",
          "warnings"
        ],
        "title": "HistoricalPage"
      },
      "HistoricalRecord": {
        "properties": {
          "log_id": {
            "type": "string",
            "title": "Log Id"
          },
          "hardware_id": {
            "type": "string",
            "title": "Hardware Id"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "cpu_utilization_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cpu Utilization Pct"
          },
          "ram_utilization_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ram Utilization Pct"
          },
          "temperatura_celsius": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperatura Celsius"
          },
          "energia_watts": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Energia Watts"
          },
          "origin": {
            "type": "string",
            "const": "unknown",
            "title": "Origin"
          },
          "source": {
            "type": "string",
            "const": "supabase",
            "title": "Source"
          },
          "quality": {
            "additionalProperties": {
              "type": "string",
              "enum": [
                "valid",
                "missing",
                "non_finite",
                "out_of_range"
              ]
            },
            "type": "object",
            "title": "Quality"
          }
        },
        "type": "object",
        "required": [
          "log_id",
          "hardware_id",
          "timestamp",
          "cpu_utilization_pct",
          "ram_utilization_pct",
          "temperatura_celsius",
          "energia_watts",
          "origin",
          "source",
          "quality"
        ],
        "title": "HistoricalRecord"
      },
...
      "Period": {
        "properties": {
          "start": {
            "type": "string",
            "format": "date-time",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "title": "End"
          }
        },
        "type": "object",
        "required": [
          "start",
          "end"
        ],
        "title": "Period"
      },
      "PredictionDataset": {
        "properties": {
          "schemaVersion": {
            "type": "string",
            "const": "1.0",
            "title": "Schemaversion"
          },
          "contractStatus": {
            "type": "string",
            "const": "provisional",
            "title": "Contractstatus"
          },
          "datasetId": {
            "type": "string",
            "title": "Datasetid"
          },
          "requestedPeriod": {
            "$ref": "#/components/schemas/Period"
          },
          "period": {
            "$ref": "#/components/schemas/Period"
          },
          "resource": {
            "$ref": "#/components/schemas/Resource"
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/PredictionFeature"
            },
            "type": "array",
            "title": "Features"
          },
          "units": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Units"
          },
          "origins": {
            "items": {
              "type": "string",
              "enum": [
                "observed",
                "simulated",
                "estimated",
                "unknown"
              ]
            },
            "type": "array",
            "title": "Origins"
          },
          "dataStatus": {
            "type": "string",
            "enum": [
              "complete",
              "partial",
              "no_data"
            ],
            "title": "Datastatus"
          },
          "excludedSampleCount": {
            "type": "integer",
            "title": "Excludedsamplecount"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "schemaVersion",
          "contractStatus",
          "datasetId",
          "requestedPeriod",
          "period",
          "resource",
          "features",
          "units",
          "origins",
          "dataStatus",
          "excludedSampleCount",
          "warnings"
        ],
        "title": "PredictionDataset"
      },
      "PredictionFeature": {
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "cpu_utilization": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Cpu Utilization"
          },
          "origin": {
            "type": "string",
            "enum": [
              "observed",
              "simulated",
              "estimated",
              "unknown"
            ],
            "title": "Origin"
          },
          "quality": {
            "type": "string",
            "const": "ok",
            "title": "Quality"
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "cpu_utilization",
          "origin",
          "quality"
        ],
        "title": "PredictionFeature"
      },
...
      "Resource": {
        "properties": {
          "type": {
            "type": "string",
            "const": "node",
            "title": "Type"
          },
          "cluster": {
            "type": "string",
            "title": "Cluster"
          },
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "cluster",
          "id"
        ],
        "title": "Resource"
      }
    }
  }
}